requirements.txt
streamlit==1.32.0
langchain==0.1.17
langchain-core==0.1.52
langchain-community==0.0.38
langchain-text-splitters==0.0.1
langchain-google-genai==1.0.3
faiss-cpu==1.7.4
sentence-transformers==2.6.1
python-dotenv==1.0.1
pypdf==4.2.0
python-docx==1.1.0
tavily-python==0.3.3